+++ doc/changelog.rst
@@ -14,6 +14,7 @@ Description
 
 - Improve documentation of exception handling in Redis server classes
 - Improve error handling for setting of scripts and models
+- Split dataset prefixing control from use_tensor_ensemble_prefix() to use_dataset_ensemble_prefix()
 - Update to the latest version of PyBind
 - Change documentation theme to sphinx_book_theme and fix doc strings
 - Add print capability for Client and DataSet
@@ -33,6 +34,7 @@ Detailed Notes
 
 - Added throw documentation for exception handling in redis.h, redisserver.h, rediscluster.h (PR301_)
 - Added error handling for a rare edge condition when setting scripts and models (PR300_)
+- The use_tensor_ensemble_prefix() API method no longer controls whether datasets are prefixed. A new API method, use_dataset_ensemble_prefix() now manages this. (PR298_)
 - Updated from PyBind v2.6.2 to v2.10.3 (PR295_)
 - Change documentation theme to sphinx_book_theme to match SmartSim documentation theme and fix Python API doc string errors (PR294_)
 - Added print capability for Client and DataSet to give details diagnostic information for debugging (PR293_)
@@ -54,9 +56,10 @@ Detailed Notes
 
 .. _PR301: https://github.com/CrayLabs/SmartRedis/pull/301
 .. _PR300: https://github.com/CrayLabs/SmartRedis/pull/300
+.. _PR298: https://github.com/CrayLabs/SmartRedis/pull/298
 .. _PR295: https://github.com/CrayLabs/SmartRedis/pull/295
 .. _PR294: https://github.com/CrayLabs/SmartRedis/pull/294
-.. _PR294: https://github.com/CrayLabs/SmartRedis/pull/293
+.. _PR293: https://github.com/CrayLabs/SmartRedis/pull/293
 .. _PR291: https://github.com/CrayLabs/SmartRedis/pull/291
 .. _PR289: https://github.com/CrayLabs/SmartRedis/pull/289
 .. _PR288: https://github.com/CrayLabs/SmartRedis/pull/288